🐛 Use Station ID for topic name
The 'name' attribute is not unique, many stations can have the same name
and even name + road number combination. Since you can run multiple
copies of trafikvader for different stations we need to ensure the topic
name is truly unique. The only way to do that is to use the ID, which
unfortunately results in humanly meaningless topic names.