🐛 Use TotalWaterEquivalent
Previously the precipitation reported could be of any value, rain, snow,
hail etc. But that makes the data incomparable since 1mm of rain is not
the same thing as 1mm of snow. TotalWaterEquivalent fixes that issue,
by always giving us a value that is "as water", so the value between
snow and water becomes comparable.

Since the original value in the Measurement on a WeatherStation was in
mm per hour, this takes the mm/30min and multiplies it by 2.